move logic to 'meetups' package
